show canvas resizer only in preview mode of anvil
